refactor: minor refactor made for some UI molecules
