+++ src/components/UI/molecules/HeaderTabs/HeaderTabs.styles.ts
@@ -4,7 +4,7 @@ import { IHeaderTabs } from './HeaderTabs.interface'
 const addDivider = ({ addDivider = false }: IHeaderTabs) => {
   if (addDivider) {
     return css`
-      border-right: 1px solid #e9e9f6;
+      border-right: 1px solid ${(props) => props.theme.colors.disabledColors.gray2};
       margin-right: 8px;
       padding-right: 8px;
     `
