+++ src/components/UI/molecules/MobileSearchbar/MobileSearchbar.component.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useRef, useState } from 'react'
+import { MainButton } from '../../atoms'
 import { IMobileSearchbar } from './MobileSearchbar.interface'
 import { withStyles } from './MobileSearchbar.styles'
-import { MainButton } from '../../atoms/MainButton'
 
 const Component: React.FC<IMobileSearchbar> = ({
   searchProps,
