+++ src/components/UI/molecules/MobileSearchbar/MobileSearchbar.interface.ts
@@ -1,4 +1,4 @@
-import { IMainButton } from '../../atoms/MainButton'
+import { IMainButton } from '../../atoms'
 
 export interface IMobileSearchbar {
   /**
