+++ src/components/UI/molecules/Searchbar/Searchbar.interface.ts
@@ -1,4 +1,4 @@
-import { IMainButton } from '../../atoms/MainButton'
+import { IMainButton } from '../../atoms'
 
 export interface ISearchbar {
   /**
