+++ src/components/UI/molecules/TabButton/TabButton.component.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { ITabButton } from './TabButton.interface'
 import { withStyles } from './TabButton.styles'
-import { Link } from '../../atoms/Link'
+import { Link } from '../../atoms'
 
 const Component: React.FC<ITabButton> = ({ boxShadow = true, linkProps, className }) => {
   return (
